fix: 프로필 사진 페이지 제거
