+++ 14th-team5-iOS/App/Sources/Presentation/Account/AccountSignUp/ViewControllers/AccountProfileViewController.swift
@@ -37,11 +37,11 @@ final class AccountProfileViewController: BaseViewController<AccountSignUpReacto
     }
     
     private func bindInput(reactor: AccountSignUpReactor) {
-        nextButton.rx.tap
-            .throttle(RxConst.throttleInterval, scheduler: Schedulers.main)
-            .map { Reactor.Action.didTapCompletehButton }
-            .bind(to: reactor.action)
-            .disposed(by: disposeBag)
+//        nextButton.rx.tap
+//            .throttle(RxConst.throttleInterval, scheduler: Schedulers.main)
+//            .map { Reactor.Action.didTapCompletehButton }
+//            .bind(to: reactor.action)
+//            .disposed(by: disposeBag)
         
         profileButton.rx.tap
             .throttle(RxConst.throttleInterval, scheduler: Schedulers.main)
@@ -67,11 +67,11 @@ final class AccountProfileViewController: BaseViewController<AccountSignUpReacto
             .bind(onNext: { $0.0.setProfilewView(with: $0.1) })
             .disposed(by: disposeBag)
         
-        reactor.state.map { $0.didTapCompletehButtonFinish }
-            .withUnretained(self)
-            .observe(on: Schedulers.main)
-            .bind(onNext: { $0.0.showNextPage(accessToken: $0.1) })
-            .disposed(by: disposeBag)
+//        reactor.state.map { $0.didTapCompletehButtonFinish }
+//            .withUnretained(self)
+//            .observe(on: Schedulers.main)
+//            .bind(onNext: { $0.0.showNextPage(accessToken: $0.1) })
+//            .disposed(by: disposeBag)
     }
     
     override func setupUI() {
@@ -151,19 +151,19 @@ extension AccountProfileViewController {
         }
     }
     
-    private func showNextPage(accessToken: AccessTokenResponse?) {
-        
-        guard let accessToken = accessToken else { return }
-        
-        let token = accessToken.accessToken
-        let refreshToken = accessToken.refreshToken
-        let isTemporaryToken = accessToken.isTemporaryToken
-        
-        let tk = AccessToken(accessToken: token, refreshToken: refreshToken, isTemporaryToken: isTemporaryToken)
-        App.Repository.token.accessToken.accept(tk)
-        
-        let container = UINavigationController(rootViewController: OnBoardingDIContainer().makeViewController())
-        container.modalPresentationStyle = .fullScreen
-        present(container, animated: false)
-    }
+//    private func showNextPage(accessToken: AccessTokenResponse?) {
+//        
+//        guard let accessToken = accessToken else { return }
+//        
+//        let token = accessToken.accessToken
+//        let refreshToken = accessToken.refreshToken
+//        let isTemporaryToken = accessToken.isTemporaryToken
+//        
+//        let tk = AccessToken(accessToken: token, refreshToken: refreshToken, isTemporaryToken: isTemporaryToken)
+//        App.Repository.token.accessToken.accept(tk)
+//        
+//        let container = UINavigationController(rootViewController: OnBoardingDIContainer().makeViewController())
+//        container.modalPresentationStyle = .fullScreen
+//        present(container, animated: false)
+//    }
 }
